+++ resources/views/pages/master/gl_account_fc/action.blade.php
@@ -1,7 +1,14 @@
-<button type="button" class="btn bg-info-transparent" title="detail" data-bs-toggle="modal" data-bs-target="{{__('#modal_detail'.$model->gl_account_fc)}}"><i class="fe fe-info"></i></button>
-<a class="btn bg-warning-transparent" title="edit" data-bs-toggle="modal" data-bs-target="{{__('#modal_edit'.$model->gl_account_fc)}}"><i class="fe fe-edit"></i></a>
-<a class="btn bg-danger-transparent" onclick="delete_gl_account_fc('{{$model->gl_account_fc}}')" title="hapus" data-toggle="tooltip"><i class="fe fe fe-trash"></i></a>
+@if (mapping_akses('gl_account_fc','read'))
+    <button type="button" class="btn bg-info-transparent" title="detail" data-bs-toggle="modal" data-bs-target="{{__('#modal_detail'.$model->gl_account_fc)}}"><i class="fe fe-info"></i></button>
+@endif
 
+@if (mapping_akses('gl_account_fc','update'))
+    <a class="btn bg-warning-transparent" title="edit" data-bs-toggle="modal" data-bs-target="{{__('#modal_edit'.$model->gl_account_fc)}}"><i class="fe fe-edit"></i></a>
+@endif
+
+@if (mapping_akses('gl_account_fc','delete'))
+    <a class="btn bg-danger-transparent" onclick="delete_gl_account_fc('{{$model->gl_account_fc}}')" title="hapus" data-toggle="tooltip"><i class="fe fe fe-trash"></i></a>
+@endif
 
 <!-- Modal Detail-->
 <div class="modal fade" id="{{__('modal_detail'.$model->gl_account_fc)}}" data-bs-backdrop="static" data-bs-keyboard="false" role="dialog" aria-labelledby="modal_detail"
